requirements/discord-handlers.txt
flask==2.3.3
gunicorn==21.2.0
python-dotenv==1.0.0
requests==2.31.0
discord.py==2.3.2
aiohttp==3.9.0 